+++ RDBqc.Rcheck/00_pkg_src/RDBqc/R/MEDBS_comp_land_Q_VL.R
@@ -26,16 +26,22 @@
 MEDBS_comp_land_Q_VL <- function(data, SP, MS, GSA,verbose=TRUE) {
   if (FALSE) {
     MS <- "ITA"
-    GSA <- "GSA 9"
-    SP <- "DPS"
-    # verbose=TRUE
-    data <- Landing_tab_example
-    MEDBS_comp_land_Q_VL(data = Landing_tab_example, MS = "ITA", GSA = "GSA 9", SP = "DPS")
+    GSA <- "GSA 18"
+    SP <- "ANE"
+    verbose=TRUE
+    data <- Land # Landing_tab_example
+    MEDBS_comp_land_Q_VL(data = Land, MS = "ITA", GSA = "GSA 18", SP = "ANE")
   }
 
   GEAR <- LANDINGS <- QUARTER <- VESSEL_LENGTH <- YEAR <- tmp1 <- tot_VL <- tot_NoVL <- NULL
 
   colnames(data) <- toupper(colnames(data))
+  data[is.na(data$VESSEL_LENGTH),"VESSEL_LENGTH"] <- "NA"
+  data[is.na(data$GEAR),"GEAR"] <- "NA"
+  data[is.na(data$MESH_SIZE_RANGE),"MESH_SIZE_RANGE"] <- "NA"
+  data[is.na(data$FISHERY),"FISHERY"] <- "NA"
+
+
   land <- data
   # land$area <- as.numeric(gsub("[^0-9.-]+", "\\1", land$area))
   land <- land[which(land$AREA == as.character(GSA) & land$COUNTRY == MS & land$SPECIES == SP), ]